refactor(chrome-extension): split common logic
